Aadded refetch for filter
