+++ importer-ui/src/features/review/index.tsx
@@ -22,7 +22,6 @@ export default function Review({ onCancel, onSuccess, upload, showImportLoadingS
 
   const { data, error, isLoading }: any = useReview(uploadId);
   const csvData = data?.data?.rows || [];
-  const [columnDefs, setColumnDefs] = useState<any>([]);
 
   const theme = useThemeStore((state) => state.theme);
 
@@ -31,54 +30,9 @@ export default function Review({ onCancel, onSuccess, upload, showImportLoadingS
 
   const isStored = data?.is_stored || {};
 
-  const defaultColDef = useMemo(() => {
-    return { sortable: true, filter: true, resizable: true };
-  }, []);
-
   const cellClickedListener = useCallback((event: any) => {
     console.log("cellClicked", event);
   }, []);
-
-  useEffect(() => {
-    if (csvData.length > 0) {
-      const headers = Object.keys(csvData[0].values);
-      const generatedColumnDefs = headers.map((header: string) => {
-        return {
-          headerName: header,
-          field: `values.${header}`,
-          cellStyle: (params: any) => {
-            if (params.data?.errors?.[header]) {
-              return { backgroundColor: "#f04339" };
-            }
-            return null;
-          },
-          cellRenderer: (params: ICellRendererParams) => {
-            if (params.data) {
-              return (
-                <span
-                  style={{
-                    display: "flex",
-                    justifyContent: "space-between",
-                    width: "100%",
-                  }}>
-                  <span>{params.value}</span>
-                  {params.data?.errors?.[header] && (
-                    <button title={params.data.errors[header][0].message}>
-                      <Icon icon="help" />
-                    </button>
-                  )}
-                </span>
-              );
-            }
-          },
-          sortable: false,
-          filter: false,
-          suppressMovable: true,
-        } as ColDef;
-      });
-      setColumnDefs(generatedColumnDefs.reverse());
-    }
-  }, [csvData]);
 
   // useEffect(() => {
   //   if (isStored || error) {
@@ -112,13 +66,7 @@ export default function Review({ onCancel, onSuccess, upload, showImportLoadingS
         <div className={style.reviewContainer}>
           <ToggleFilter options={filterOptions} className={style.filters} onChange={(option: string) => onFilterChange(option)} />
           <div className={style.tableWrapper}>
-            <ReviewDataTable
-              columnDefs={columnDefs}
-              defaultColDef={defaultColDef}
-              cellClickedListener={cellClickedListener}
-              theme={theme}
-              uploadId={uploadId}
-            />
+            <ReviewDataTable cellClickedListener={cellClickedListener} theme={theme} uploadId={uploadId} filter={filter} />
           </div>
           <div className={style.actions}>
             <Button type="button" variants={["secondary"]} onClick={onCancel}>
